Feature/scrum 140: Estandarizacion de funciones CRUD (#70)
Se movieron los files necesarios para CRUD divisa y tasa de cambio, a panel_admin.

• Views, templates y urls que son necesarios par realizar CRUD se sacaron de app operaciones
• Los test que evaluan las fuciones para CRUD tambien se movieron de la app operaciones a la app panel_admin
