+++ app/apps/operaciones/tests/test_tasa_cambio_crear.py
@@ -30,7 +30,7 @@ class TasaCambioCrearTest(TestCase):
             "activo": True,
         }
 
-        response = self.client.post(reverse("operaciones:crear_tasa"), data)
+        response = self.client.post(reverse("tasa_cambio_crear"), data)
 
         # Verificar que la respuesta sea exitosa (302 = redirección o 200 = OK)
         self.assertIn(response.status_code, [200, 302])
@@ -51,7 +51,7 @@ class TasaCambioCrearTest(TestCase):
             "activo": True,
         }
 
-        response = self.client.post(reverse("operaciones:crear_tasa"), data)
+        response = self.client.post(reverse("tasa_cambio_crear"), data)
 
         # Debe retornar 200 (formulario con errores) no 302 (redirección exitosa)
         self.assertEqual(response.status_code, 200)
