+++ app/apps/operaciones/tests/test_tasa_cambio_listar.py
@@ -34,6 +34,6 @@ class TasaCambioListarTest(TestCase):
         )
 
     def test_tasa_cambio_listar(self):
-        response = self.client.get(reverse("operaciones:tasa_cambio_listar"))
+        response = self.client.get(reverse("tasa_cambio_listar"))
         self.assertEqual(response.status_code, 200)
         self.assertIn("tasas_de_cambio", response.context)
