+++ app/apps/panel_admin/templates/divisa_detalle.html
@@ -7,12 +7,11 @@
 {% block content %}
     <div class="container mx-auto px-4 py-8">
         <div class="mb-4">
-            <a href="{% url 'operaciones:divisa_list' %}"
-               class="btn btn-sm btn-outline">← Volver a la lista</a>
+            <a href="{% url 'divisa_list' %}" class="btn btn-sm btn-outline">← Volver a la lista</a>
         </div>
         <div class="flex justify-between items-center mb-6">
             <h1 class="text-2xl font-bold text-blue-800">Detalle Divisa</h1>
-            <a href="{% url 'operaciones:crear_divisa' %}" class="btn btn-primary">
+            <a href="{% url 'crear_divisa' %}" class="btn btn-primary">
                 <svg xmlns="http://www.w3.org/2000/svg"
                      class="h-5 w-5 mr-2"
                      fill="none"
@@ -23,7 +22,6 @@
                 Crear Divisa
             </a>
         </div>
-        
         <div class="card w-full max-w-2xl mx-auto bg-white shadow-xl overflow-auto rounded-lg">
             <div class="card-body">
                 <h2 class="text-xl font-semibold mb-4">{{ divisa.nombre }}</h2>
@@ -38,10 +36,9 @@
                     </div>
                 </div>
                 <div class="mt-6 flex space-x-2">
-                    <a href="{% url 'operaciones:edit_divisa' divisa.pk %}"
-                       class="btn btn-primary">Editar</a>
+                    <a href="{% url 'edit_divisa' divisa.pk %}" class="btn btn-primary">Editar</a>
                     <form method="POST"
-                          action="{% url 'operaciones:delete_divisa' divisa.pk %}"
+                          action="{% url 'delete_divisa' divisa.pk %}"
                           class="inline">
                         {% csrf_token %}
                         <button type="button"
@@ -58,8 +55,7 @@
                 <h3 class="font-bold text-lg">Confirmar Eliminación</h3>
                 <p class="py-4">¿Estás seguro de que deseas eliminar la divisa "{{ divisa.nombre }}"?</p>
                 <div class="modal-action mt-4 flex justify-end space-x-2">
-                    <form method="POST"
-                          action="{% url 'operaciones:delete_divisa' divisa.pk %}">
+                    <form method="POST" action="{% url 'delete_divisa' divisa.pk %}">
                         {% csrf_token %}
                         <button type="submit" class="btn btn-error">Eliminar</button>
                     </form>
