+++ app/apps/panel_admin/templates/divisa_list.html
@@ -78,7 +78,7 @@
                                     </form>
                                     <h3 class="font-bold text-lg text-blue-700 mb-4">Editar Divisa</h3>
                                     <form method="post"
-                                          action="{% url 'operaciones:edit_divisa' divisa.pk %}"
+                                          action="{% url 'edit_divisa' divisa.pk %}"
                                           class="space-y-4">
                                         {% csrf_token %}
                                         <div class="grid grid-cols-2 gap-4">
@@ -149,7 +149,7 @@
                                         ¿Estás seguro que deseas eliminar la divisa <span class="font-semibold">{{ divisa.codigo }} - {{ divisa.nombre }}</span>?
                                     </p>
                                     <form method="post"
-                                          action="{% url 'operaciones:delete_divisa' divisa.pk %}">
+                                          action="{% url 'delete_divisa' divisa.pk %}">
                                         {% csrf_token %}
                                         <div class="modal-action mt-4 flex justify-end space-x-2">
                                             <button type="submit" class="btn btn-error">Eliminar</button>
@@ -189,7 +189,7 @@
             <h3 class="font-bold text-lg text-blue-800 mb-4">Crear Divisa</h3>
             <form id="crear-divisa-form"
                   method="POST"
-                  action="{% url 'operaciones:crear_divisa' %}">
+                  action="{% url 'crear_divisa' %}">
                 {% csrf_token %}
                 <div class="form-control w-full mb-4 relative">
                     <label class="label">
@@ -201,7 +201,8 @@
                            required
                            maxlength="3"
                            class="input input-bordered w-full"
-                           placeholder="Ej. USD, EUR, PYG">
+                           placeholder="Ej. USD, EUR, PYG"
+                           autocomplete="off">
                     <ul id="suggestions-modal"
                         class="bg-white border border-gray-300 rounded-lg shadow-lg max-h-48 overflow-y-auto z-10 hidden absolute w-full mt-2">
                     </ul>
@@ -223,8 +224,8 @@
                     <input type="text"
                            name="simbolo"
                            id="id_simbolo_modal"
-                           readonly
-                           class="input input-bordered w-full bg-gray-100 cursor-not-allowed">
+                           class="input input-bordered w-full"
+                           placeholder="Simbolo opcional">
                 </div>
                 <div class="form-control w-full mb-4">
                     <label class="label">
@@ -278,7 +279,7 @@
 
                 const fetchCurrencies = async () => {
                     try {
-                        const apiUrl = "/operaciones/admin/divisas/api/";
+                        const apiUrl = "/admin/divisas/api/";
                         const response = await fetch(apiUrl);
                         if (!response.ok) throw new Error('Network response was not ok');
 
@@ -345,7 +346,7 @@
                     const formData = new FormData(form);
 
                     try {
-                        const response = await fetch("{% url 'operaciones:crear_divisa' %}", {
+                        const response = await fetch("{% url 'crear_divisa' %}", {
                             method: 'POST',
                             body: formData,
                             headers: {
