+++ app/apps/panel_admin/templates/panel_inicio.html
@@ -32,15 +32,15 @@
                                        class="btn btn-outline btn-accent w-full my-2">Clientes</a>
                                 </li>
                                 <li>
-                                    <a href="{% url 'operaciones:divisa_list' %}"
+                                    <a href="{% url 'divisa_list' %}"
                                        class="btn btn-outline btn-info w-full my-2">Divisas</a>
                                 </li>
                                 <li>
-                                    <a href="{% url 'operaciones:tasa_cambio_listar' %}"
+                                    <a href="{% url 'tasa_cambio_listar' %}"
                                        class="btn btn-outline btn-ghost w-full my-2">Tasas de Cambio</a>
                                 </li>
                                 <li>
-                                    <a href="{% url 'operaciones:tasa_cambio_historial_listar' %}"
+                                    <a href="{% url 'tasa_cambio_historial_listar' %}"
                                        class="btn btn-outline btn-warning w-full my-2">Historial de Tasas de Cambio</a>
                                 </li>
                             </ul>
