+++ app/apps/panel_admin/templates/tasa_cambio_form.html
@@ -93,7 +93,7 @@
                         {% endfor %}
                     </div>
                     <div class="mt-6 flex justify-end space-x-4">
-                        <a href="{% url 'operaciones:tasa_cambio_listar' %}"
+                        <a href="{% url 'tasa_cambio_listar' %}"
                            class="btn btn-outline btn-neutral">Cancelar</a>
                         <button type="submit" class="btn btn-primary">
                             {% if form.instance.pk %}
