+++ app/apps/panel_admin/tests/test_delete_divisa.py
@@ -11,13 +11,13 @@ class DivisaViewsTestCase(TestCase):
         self.divisa = Divisa.objects.create(**self.divisa_data)
 
     def test_delete_divisa_post(self):
-        response = self.client.post(reverse("operaciones:delete_divisa", args=[self.divisa.pk]))
+        response = self.client.post(reverse("delete_divisa", args=[self.divisa.pk]))
 
         self.assertEqual(response.status_code, 302)
         self.assertFalse(Divisa.objects.filter(pk=self.divisa.pk).exists())
 
     def test_delete_divisa_get_redirects(self):
-        response = self.client.get(reverse("operaciones:delete_divisa", args=[self.divisa.pk]))
+        response = self.client.get(reverse("delete_divisa", args=[self.divisa.pk]))
 
         self.assertEqual(response.status_code, 302)
         self.assertTrue(Divisa.objects.filter(pk=self.divisa.pk).exists())
